question50
At first time, I start the iteration from the smallest length of the consecutive prime number which turns out to be highly cost about 17 s. 
Then I reverse the order of iteration starting from the largest length of the consecutive prime number which is determined automatically, the cost is reduced within a second.